requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
pydantic-ai>=0.0.12
SQLAlchemy>=2.0.23
alembic>=1.12.1
qdrant-client>=1.6.4
redis>=5.0.1
confluent-kafka>=2.3.0
websockets>=12.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
requests>=2.31.0
PyYAML>=6.0.1
